uses float for number of shares

+++ trades/models.py
@@ -37,7 +37,7 @@ class Trade(models.Model):
     currency = models.CharField('Currency', max_length=3, default='USD', choices=CURRENCIES)
     action = models.CharField('Action', max_length=4, default='BUY', choices=ACTIONS)
     trade_date = models.DateTimeField('Date', default=timezone.now)
-    number_of_shares = models.IntegerField('Number of shares', default=1)
+    number_of_shares = models.FloatField('Number of shares', default=1)
     price_paid = models.FloatField('Price paid')
     current_price = models.FloatField('Last close value', default=0)
     total_value = models.FloatField('Total value', default=0)
